requirements.txt
selenium
requests
pycryptodome
pillow
jinja2
pyyaml